refactor: delete dead code and rationalize code
