+++ packages/core/src/data-source/mutable-data-source.ts
@@ -19,8 +19,8 @@ import {
   stringToSortableHashes,
   applySort,
   HASH_CHUNK_COUNT,
-} from "./sorting";
-import { applyFilters } from "./filtering";
+} from "../indexed-data-store/sorting";
+import { applyFilters } from "../filtering";
 import { createInstructionEmitter } from "../utils";
 
 // =============================================================================
