--- packages/core/src/indexed-data-store/filtering.ts
@@ -1,12 +1,0 @@
-// packages/core/src/indexed-data-store/filtering.ts
-// Re-exports from shared filtering module
-
-export {
-  isSameDay,
-  evaluateTextCondition,
-  evaluateNumberCondition,
-  evaluateDateCondition,
-  evaluateCondition,
-  evaluateColumnFilter,
-  rowPassesFilter,
-} from "../filtering";
