--- packages/core/src/types.ts
@@ -1,88 +1,0 @@
-// packages/core/src/types.ts
-// Re-export all types from modular structure for backwards compatibility
-
-export type {
-  // Basic types
-  CellDataType,
-  CellValue,
-  Row,
-  RowId,
-  SortDirection,
-  SortModel,
-  CellPosition,
-  CellRange,
-  SelectionState,
-  EditState,
-  FillHandleState,
-  CellValueChangedEvent,
-  SlotState,
-  // Column types
-  ColumnDefinition,
-  // Filter types
-  TextFilterOperator,
-  NumberFilterOperator,
-  DateFilterOperator,
-  FilterCombination,
-  TextFilterCondition,
-  NumberFilterCondition,
-  DateFilterCondition,
-  FilterCondition,
-  ColumnFilterModel,
-  FilterModel,
-  // Data source types
-  DataSourceRequest,
-  DataSourceResponse,
-  DataSource,
-  // Instruction types
-  CreateSlotInstruction,
-  DestroySlotInstruction,
-  AssignSlotInstruction,
-  MoveSlotInstruction,
-  SetActiveCellInstruction,
-  SetSelectionRangeInstruction,
-  StartEditInstruction,
-  StopEditInstruction,
-  CommitEditInstruction,
-  SetContentSizeInstruction,
-  UpdateHeaderInstruction,
-  OpenFilterPopupInstruction,
-  CloseFilterPopupInstruction,
-  StartFillInstruction,
-  UpdateFillInstruction,
-  CommitFillInstruction,
-  CancelFillInstruction,
-  DataLoadingInstruction,
-  DataLoadedInstruction,
-  DataErrorInstruction,
-  RowsAddedInstruction,
-  RowsRemovedInstruction,
-  RowsUpdatedInstruction,
-  TransactionProcessedInstruction,
-  SetHoverPositionInstruction,
-  ColumnsChangedInstruction,
-  StartColumnResizeInstruction,
-  UpdateColumnResizeInstruction,
-  CommitColumnResizeInstruction,
-  CancelColumnResizeInstruction,
-  StartColumnMoveInstruction,
-  UpdateColumnMoveInstruction,
-  CommitColumnMoveInstruction,
-  CancelColumnMoveInstruction,
-  StartRowDragInstruction,
-  UpdateRowDragInstruction,
-  CommitRowDragInstruction,
-  CancelRowDragInstruction,
-  GridInstruction,
-  InstructionListener,
-  BatchInstructionListener,
-  // Highlighting types
-  HighlightColumnInfo,
-  HighlightContext,
-  HighlightingOptions,
-  // Renderer types
-  CellRendererParams,
-  EditRendererParams,
-  HeaderRendererParams,
-  // Options types
-  GridCoreOptions,
-} from "./types/index";
