+++ packages/react/src/gridState/reducer.ts
@@ -1,175 +1,11 @@
 // packages/react/src/gridState/reducer.ts
 
-import type { GridInstruction, Row } from "@gp-grid/core";
-import type { SlotData, HeaderData, GridState, GridAction } from "./types";
+import type { Row } from "@gp-grid/core";
+import { createInitialState, applyInstruction } from "@gp-grid/core";
+import type { GridState, GridAction } from "./types";
 
-// =============================================================================
-// Initial State
-// =============================================================================
-
-export interface InitialStateArgs {
-  initialWidth?: number;
-  initialHeight?: number;
-}
-
-export function createInitialState<TData = Row>(args?: InitialStateArgs): GridState<TData> {
-  return {
-    slots: new Map(),
-    activeCell: null,
-    selectionRange: null,
-    editingCell: null,
-    contentWidth: 0,
-    contentHeight: args?.initialHeight ?? 0,
-    viewportWidth: args?.initialWidth ?? 0,
-    viewportHeight: args?.initialHeight ?? 0,
-    rowsWrapperOffset: 0,
-    headers: new Map(),
-    filterPopup: null,
-    isLoading: false,
-    error: null,
-    totalRows: 0,
-    visibleRowRange: null,
-    hoverPosition: null,
-    columns: null,
-  };
-}
-
-// =============================================================================
-// Instruction Handler
-// =============================================================================
-
-/**
- * Apply a single instruction to mutable slot maps and return other state changes.
- * This allows batching multiple slot operations efficiently.
- */
-export function applyInstruction<TData = Row>(
-  instruction: GridInstruction,
-  slots: Map<string, SlotData<TData>>,
-  headers: Map<number, HeaderData>,
-): Partial<GridState<TData>> | null {
-  switch (instruction.type) {
-    case "CREATE_SLOT":
-      slots.set(instruction.slotId, {
-        slotId: instruction.slotId,
-        rowIndex: -1,
-        rowData: {} as TData,
-        translateY: 0,
-      });
-      return null; // Slots map is mutated
-
-    case "DESTROY_SLOT":
-      slots.delete(instruction.slotId);
-      return null;
-
-    case "ASSIGN_SLOT": {
-      const existing = slots.get(instruction.slotId);
-      if (existing) {
-        slots.set(instruction.slotId, {
-          ...existing,
-          rowIndex: instruction.rowIndex,
-          rowData: instruction.rowData as TData,
-        });
-      }
-      return null;
-    }
-
-    case "MOVE_SLOT": {
-      const existing = slots.get(instruction.slotId);
-      if (existing) {
-        slots.set(instruction.slotId, {
-          ...existing,
-          translateY: instruction.translateY,
-        });
-      }
-      return null;
-    }
-
-    case "SET_ACTIVE_CELL":
-      return { activeCell: instruction.position };
-
-    case "SET_SELECTION_RANGE":
-      return { selectionRange: instruction.range };
-
-    case "UPDATE_VISIBLE_RANGE":
-      return {
-        visibleRowRange: { start: instruction.start, end: instruction.end },
-        rowsWrapperOffset: instruction.rowsWrapperOffset,
-      };
-
-    case "SET_HOVER_POSITION":
-      return { hoverPosition: instruction.position };
-
-    case "START_EDIT":
-      return {
-        editingCell: {
-          row: instruction.row,
-          col: instruction.col,
-          initialValue: instruction.initialValue,
-        },
-      };
-
-    case "STOP_EDIT":
-      return { editingCell: null };
-
-    case "SET_CONTENT_SIZE":
-      return {
-        contentWidth: instruction.width,
-        contentHeight: instruction.height,
-        viewportWidth: instruction.viewportWidth,
-        viewportHeight: instruction.viewportHeight,
-        rowsWrapperOffset: instruction.rowsWrapperOffset,
-      };
-
-    case "UPDATE_HEADER":
-      headers.set(instruction.colIndex, {
-        column: instruction.column,
-        sortDirection: instruction.sortDirection,
-        sortIndex: instruction.sortIndex,
-        hasFilter: instruction.hasFilter,
-      });
-      return null;
-
-    case "OPEN_FILTER_POPUP":
-      return {
-        filterPopup: {
-          isOpen: true,
-          colIndex: instruction.colIndex,
-          column: instruction.column,
-          anchorRect: instruction.anchorRect,
-          distinctValues: instruction.distinctValues,
-          currentFilter: instruction.currentFilter,
-        },
-      };
-
-    case "CLOSE_FILTER_POPUP":
-      return { filterPopup: null };
-
-    case "DATA_LOADING":
-      return { isLoading: true, error: null };
-
-    case "DATA_LOADED":
-      return { isLoading: false, totalRows: instruction.totalRows };
-
-    case "DATA_ERROR":
-      return { isLoading: false, error: instruction.error };
-
-    // Transaction instructions
-    case "ROWS_ADDED":
-    case "ROWS_REMOVED":
-      return { totalRows: instruction.totalRows };
-
-    case "ROWS_UPDATED":
-    case "TRANSACTION_PROCESSED":
-      // These don't change state directly - slot updates come via ASSIGN_SLOT
-      return null;
-
-    case "COLUMNS_CHANGED":
-      return { columns: instruction.columns };
-
-    default:
-      return null;
-  }
-}
+export type { InitialStateArgs } from "@gp-grid/core";
+export { createInitialState } from "@gp-grid/core";
 
 // =============================================================================
 // Reducer
