+++ packages/vue/src/composables/useInputHandler.ts
@@ -19,6 +19,7 @@ import type {
   ColumnDefinition,
   SlotData,
 } from "@gp-grid/core";
+import { scrollCellIntoView } from "@gp-grid/core";
 import { useAutoScroll } from "./useAutoScroll";
 
 // =============================================================================
@@ -57,55 +58,6 @@ export interface UseInputHandlerResult {
   handleKeyDown: (e: KeyboardEvent) => void;
   handleWheel: (e: WheelEvent, wheelDampening: number) => void;
   dragState: Ref<DragState>;
-}
-
-// =============================================================================
-// Helper Functions
-// =============================================================================
-
-/**
- * Find the slot for a given row index
- */
-function findSlotForRow(
-  slots: Map<string, SlotData>,
-  rowIndex: number,
-): SlotData | null {
-  for (const slot of slots.values()) {
-    if (slot.rowIndex === rowIndex) {
-      return slot;
-    }
-  }
-  return null;
-}
-
-/**
- * Scroll a cell into view if needed
- */
-function scrollCellIntoView<TData extends Row = Row>(
-  core: GridCore<TData>,
-  container: HTMLDivElement,
-  row: number,
-  rowHeight: number,
-  headerHeight: number,
-  slots: Map<string, SlotData>,
-): void {
-  const slot = findSlotForRow(slots, row);
-  const cellTranslateY = slot
-    ? slot.translateY
-    : headerHeight + row * rowHeight;
-  const cellViewportTop = cellTranslateY - container.scrollTop;
-  const cellViewportBottom = cellViewportTop + rowHeight;
-  const visibleTop = headerHeight;
-  const visibleBottom = container.clientHeight;
-
-  if (cellViewportTop < visibleTop) {
-    container.scrollTop = core.getScrollTopForRow(row);
-  } else if (cellViewportBottom > visibleBottom) {
-    const visibleDataHeight = container.clientHeight - headerHeight;
-    const rowsInView = Math.floor(visibleDataHeight / rowHeight);
-    const targetRow = Math.max(0, row - rowsInView + 1);
-    container.scrollTop = core.getScrollTopForRow(targetRow);
-  }
 }
 
 // =============================================================================
