+++ packages/vue/src/renderers/editRenderer.ts
@@ -4,19 +4,7 @@ import { h, createTextVNode, type VNode } from "vue";
 import type { GridCore, ColumnDefinition, Row, CellValue, EditRendererParams } from "@gp-grid/core";
 import type { VueEditRenderer } from "../types";
 import { getCellValue } from "./cellRenderer";
-
-/**
- * Ensure we always return a VNode, never a plain string
- */
-function toVNode(value: VNode | string | null | undefined): VNode {
-  if (value == null || value === "") {
-    return createTextVNode("");
-  }
-  if (typeof value === "string") {
-    return createTextVNode(value);
-  }
-  return value;
-}
+import { toVNode } from "./utils";
 
 export interface RenderEditCellOptions {
   column: ColumnDefinition;
